requirements.txt
anyio==3.7.0
certifi==2023.5.7
dotenv-python==0.0.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
jsonpath==0.82
loguru==0.7.0
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
sniffio==1.3.0
typing_extensions==4.7.1
